refactor: menu, product 분리
