+++ src/main/java/kitchenpos/product/application/facade/ProductFacade.java
@@ -1,0 +1,5 @@
+package kitchenpos.product.application.facade;
+
+public class ProductFacade {
+
+}
