+++ src/main/java/kitchenpos/product/infrastructure/persistence/JpaProductRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.infrastructure.persistence;
+package kitchenpos.product.infrastructure.persistence;
 
 import java.util.UUID;
-import kitchenpos.menu.domain.entity.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
+import kitchenpos.product.domain.entity.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
