--- src/test/java/kitchenpos/ApplicationTest.java
@@ -1,12 +1,0 @@
-package kitchenpos;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTest {
-
-    @Test
-    void contextLoads() {
-    }
-}
