--- src/test/java/kitchenpos/application/FakePurgomalumClient.java
@@ -1,20 +1,0 @@
-package kitchenpos.application;
-
-import java.util.Arrays;
-import java.util.List;
-import kitchenpos.global.infrastructure.external.PurgomalumClient;
-
-public class FakePurgomalumClient implements PurgomalumClient {
-
-    private static final List<String> profanities;
-
-    static {
-        profanities = Arrays.asList("비속어", "욕설");
-    }
-
-    @Override
-    public boolean containsProfanity(final String text) {
-        return profanities.stream()
-            .anyMatch(profanity -> text.contains(profanity));
-    }
-}
