--- src/test/java/kitchenpos/application/InMemoryProductRepository.java
@@ -1,39 +1,0 @@
-package kitchenpos.application;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-import kitchenpos.menu.domain.entity.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
-
-public class InMemoryProductRepository implements ProductRepository {
-
-    private final Map<UUID, Product> products = new HashMap<>();
-
-    @Override
-    public Product save(final Product product) {
-        products.put(product.getId(), product);
-        return product;
-    }
-
-    @Override
-    public Optional<Product> findById(final UUID id) {
-        return Optional.ofNullable(products.get(id));
-    }
-
-    @Override
-    public List<Product> findAll() {
-        return new ArrayList<>(products.values());
-    }
-
-    @Override
-    public List<Product> findAllByIdIn(final List<UUID> ids) {
-        return products.values()
-            .stream()
-            .filter(product -> ids.contains(product.getId()))
-            .toList();
-    }
-}
