--- src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -1,63 +1,0 @@
-package kitchenpos.application;
-
-import static kitchenpos.Fixtures.menuGroup;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-import java.util.List;
-import kitchenpos.menu.domain.entity.MenuGroup;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
-import kitchenpos.menu.domain.service.MenuGroupService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
-
-class MenuGroupServiceTest {
-
-    private MenuGroupRepository menuGroupRepository;
-    private MenuGroupService menuGroupService;
-
-    @BeforeEach
-    void setUp() {
-        menuGroupRepository = new InMemoryMenuGroupRepository();
-        menuGroupService = new MenuGroupService(menuGroupRepository);
-    }
-
-    @DisplayName("메뉴 그룹을 등록할 수 있다.")
-    @Test
-    void create() {
-        final MenuGroup expected = createMenuGroupRequest("두마리메뉴");
-        final MenuGroup actual = menuGroupService.create(expected);
-        assertThat(actual).isNotNull();
-        assertAll(
-            () -> assertThat(actual.getId()).isNotNull(),
-            () -> assertThat(actual.getName()).isEqualTo(expected.getName())
-        );
-    }
-
-    @DisplayName("메뉴 그룹의 이름이 올바르지 않으면 등록할 수 없다.")
-    @NullAndEmptySource
-    @ParameterizedTest
-    void create(final String name) {
-        final MenuGroup expected = createMenuGroupRequest(name);
-        assertThatThrownBy(() -> menuGroupService.create(expected))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("메뉴 그룹의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        menuGroupRepository.save(menuGroup("두마리메뉴"));
-        final List<MenuGroup> actual = menuGroupService.findAll();
-        assertThat(actual).hasSize(1);
-    }
-
-    private MenuGroup createMenuGroupRequest(final String name) {
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setName(name);
-        return menuGroup;
-    }
-}
